added new info.java file and moved howtoplay there
